+++ src/config.js
@@ -90,10 +90,10 @@ export default Object.assign(
     BUILD_VERSION: packageJson.version,
 
     /**
-     * User will be logged out after IDLE_TIMEOUT (milliseconds) time of
+     * User will be logged out after IDLE_TIMEOUT (minutes) time of
      * inactivity
      */
-    IDLE_TIMEOUT: 1000 * 60 * 15,
+    IDLE_TIMEOUT: 15,
 
     /**
      * Module scheme to use. The module scheme will be used to combine module
@@ -116,6 +116,11 @@ export default Object.assign(
      * required by the Sentry SDKs.
      */
     SENTRY_DSN: '',
+
+    /**
+     * Keep session interval
+     */
+    KEEP_SESSION_INTERVAL: 1000 * 60 * 2,
 
     /**
      * Change over different recovery modes. To disable recovery, leave the