+++ src/vue/fields/AmountInputField.vue
@@ -54,6 +54,7 @@ const AMOUNT_VALIDATION_TYPE = {
   incoming: 'incoming',
   outgoing: 'outgoing',
   issuance: 'issuance',
+  atomicSwap: 'atomicSwap',
 }
 
 export default {
@@ -132,6 +133,12 @@ export default {
 
         case AMOUNT_VALIDATION_TYPE.issuance:
           result = this.assetRecord.availableForIssuance
+          break
+
+        case AMOUNT_VALIDATION_TYPE.atomicSwap:
+          result = MathUtil.add(
+            this.assetRecord.availableForIssuance, this.balance
+          )
           break
 
         default: