+++ src/vue/modules/create-sale-form/components/information-step-form.vue
@@ -336,11 +336,12 @@ export default {
 
   computed: {
     priceForAsset () {
+      let value = MathUtil.divide(
+        this.form.hardCap,
+        this.form.assetsToSell
+      )
       return {
-        value: MathUtil.divide(
-          this.form.hardCap,
-          this.form.assetsToSell
-        ),
+        value: isNaN(value) ? 0 : value,
         currency: this.form.capAsset.code,
       }
     },
@@ -363,7 +364,9 @@ export default {
         this.form.quoteAssets = []
       }
 
-      this.availableQuoteAssets = await this.loadBaseAssetsByQuote(value)
+      const quoteAssets = await this.loadBaseAssetsByQuote(value)
+
+      this.availableQuoteAssets = [this.form.capAsset, ...quoteAssets]
       this.isQuoteAssetsLoaded = true
     },
   },
@@ -391,9 +394,6 @@ export default {
     },
 
     populateForm () {
-      const quoteAssets = this.request.quoteAssets
-        .filter(quoteAsset => this.request.defaultQuoteAsset !== quoteAsset)
-
       this.form.name = this.request.name
       this.form.baseAsset = this.ownedAssets
         .find(item => item.code === this.request.baseAsset)
@@ -404,7 +404,7 @@ export default {
       this.form.softCap = this.request.softCap
       this.form.hardCap = this.request.hardCap
       this.form.assetsToSell = this.request.assetsToSell
-      this.form.quoteAssets = quoteAssets
+      this.form.quoteAssets = this.request.quoteAssets
       this.form.isWhitelisted = this.request.isWhitelisted
     },