+++ src/vuex/types.js
@@ -17,6 +17,7 @@ const mutations = {
   SET_KV_POLL_TYPE_UNRESTRICTED: 'SET_KV_POLL_TYPE_UNRESTRICTED',
 
   SET_DEFAULT_QUOTE_ASSET: 'SET_DEFAULT_QUOTE_ASSET',
+  SET_KV_DEFAULT_SIGNER_ROLE_ID: 'SET_KV_DEFAULT_SIGNER_ROLE_ID',
 
   // account
   SET_ACCOUNT: 'SET_ACCOUNT',
@@ -24,6 +25,7 @@ const mutations = {
 
   // wallet
   SET_WALLET: 'SET_WALLET',
+  SET_ENCRYPTED_SECRET_SEED: 'SET_ENCRYPTED_SECRET_SEED',
 
   // factors
   SET_FACTORS: 'SET_FACTORS',
@@ -35,9 +37,16 @@ const mutations = {
   SET_ACCOUNT_ROLE_RESETED: 'SET_ACCOUNT_ROLE_RESETED',
   SET_KYC_LATEST_REQUEST_DATA: 'SET_KYC_LATEST_REQUEST_DATA',
 
+  // kyc recovery
+  SET_KYC_RECOVERY_LATEST_REQUEST: 'SET_KYC_RECOVERY_LATEST_REQUEST',
+  SET_KYC_RECOVERY_LATEST_REQUEST_BLOB: 'SET_KYC_RECOVERY_LATEST_REQUEST_BLOB',
+
   // assets
   SET_ASSETS: 'SET_ASSETS',
   UPDATE_ASSETS: 'UPDATE_ASSETS',
+
+  // idle
+  UPDATE_LOGOUT_AT: 'UPDATE_LOGOUT_AT',
 }
 
 const actions = {
@@ -52,6 +61,9 @@ const actions = {
   // wallet
   LOAD_WALLET: 'LOAD_WALLET',
   STORE_WALLET: 'STORE_WALLET',
+  DECRYPT_SECRET_SEED: 'DECRYPT_SECRET_SEED',
+  GET_SESSION: 'GET_SESSION',
+  PROLONGATE_SESSION: 'PROLONGATE_SESSION',
 
   // factors
   LOAD_FACTORS: 'LOAD_FACTORS',
@@ -63,8 +75,22 @@ const actions = {
   LOAD_KYC_RELATED_REQUEST: 'LOAD_KYC_RELATED_REQUEST',
   LOAD_KYC_LATEST_DATA: 'LOAD_KYC_LATEST_DATA',
 
+  // kyc recovery
+  LOAD_KYC_RECOVERY_LATEST_REQUEST: 'LOAD_KYC_RECOVERY_LATEST_REQUEST',
+  LOAD_KYC_RECOVERY: 'LOAD_KYC_RECOVERY',
+  LOAD_KYC_RECOVERY_LATEST_REQUEST_BLOB:
+    'LOAD_KYC_RECOVERY_LATEST_REQUEST_BLOB',
+  SEND_KYC_RECOVERY_REQUEST: 'SEND_KYC_RECOVERY_REQUEST',
+
   // assets
   LOAD_ASSETS: 'LOAD_ASSETS',
+
+  // idle
+  START_IDLE: 'START_IDLE',
+  LOGOUT_IDLE: 'LOGOUT_IDLE',
+  INIT_IDLE_TICKER: 'INIT_IDLE_TICKER',
+  KEEP_SESSION: 'KEEP_SESSION',
+  LOGOUT_SESSION: 'LOGOUT_SESSION',
 }
 
 const getters = {
@@ -84,6 +110,7 @@ const getters = {
   kvPollTypeRestricted: 'kvPollTypeRestricted',
   kvPollTypeUnrestricted: 'kvPollTypeUnrestricted',
   defaultQuoteAsset: 'defaultQuoteAsset',
+  kvDefaultSignerRoleId: 'kvDefaultSignerRoleId',
 
   // account
   account: 'account',
@@ -93,6 +120,7 @@ const getters = {
   accountBalanceByCode: 'accountBalanceByCode',
   accountRoleId: 'accountRoleId',
   accountDepositAddresses: 'accountDepositAddresses',
+  accountKycRecoveryStatus: 'accountKycRecoveryStatus',
 
   isAccountGeneral: 'isAccountGeneral',
   isAccountCorporate: 'isAccountCorporate',
@@ -105,7 +133,9 @@ const getters = {
   walletId: 'walletId',
   walletAccountId: 'walletAccountId',
   walletEmail: 'walletEmail',
-  walletSeed: 'walletSeed',
+  sessionId: 'sessionId',
+  encryptedSecretSeed: 'encryptedSecretSeed',
+  walletPublicKey: 'walletPublicKey',
 
   // factors
   factors: 'factors',
@@ -130,6 +160,21 @@ const getters = {
   kycLatestRequestData: 'kycLatestRequestData',
   kycAvatarKey: 'kycAvatarKey',
   isAccountRoleReseted: 'isAccountRoleReseted',
+
+  // kyc recovery
+  kycRecoveryRequestId: 'kycRecoveryRequestId',
+  kycRecoveryState: 'kycRecoveryState',
+  kycRecoveryStateI: 'kycRecoveryStateI',
+  kycRecoveryRequestBlob: 'kycRecoveryRequestBlob',
+  kycRecoveryRejectReason: 'kycRecoveryRejectReason',
+  kycRecoveryBlobId: 'kycRecoveryBlobId',
+  isNoKycRecoveryInProgress: 'isNoKycRecoveryInProgress',
+  isKycRecoveryInProgress: 'isKycRecoveryInProgress',
+  isKycRecoveryInited: 'isKycRecoveryInited',
+  isKycRecoveryApproved: 'isKycRecoveryApproved',
+  isKycRecoveryPending: 'isKycRecoveryPending',
+  isKycRecoveryRejected: 'isKycRecoveryRejected',
+  isKycRecoveryPermanentlyRejected: 'isKycRecoveryPermanentlyRejected',
 
   // assets
   assets: 'assets',